IMPALA-8954: Uncorrelated scalar subqueries in the select list
Extend StmtRewriter with the ability to rewrite scalar subqueries in the
select list into cross joins. Currently the subquery must pass plan-time
checks to determine that it returns a single row which may miss cases
that may be valid at runtime or with more complex evaluation of the
predicate expressions in the planner. Support for correlated subqueries
will be a follow on change.

Testing:
  * Added new analyzer tests, updated previous subquery tests
  * test_queries.py::TestQueries::test_subquery
  * Added test_tpcds_q9 to e2e and planner tests

Reviewed-on: http://gerrit.cloudera.org:8080/16007
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>

+++ testdata/datasets/tpcds/tpcds_schema_template.sql
@@ -1138,7 +1138,7 @@ tpcds
 ---- BASE_TABLE_NAME
 reason
 ---- COLUMNS
-r_reason_sk           bigint
+r_reason_sk           int
 r_reason_id           string
 r_reason_desc         string
 ---- ROW_FORMAT
